fix: remove and comment hermes related
