--- models/relayer/constants.go
@@ -1,15 +1,0 @@
-package relayer
-
-const (
-	HermesVersion = "v1.10.4"
-
-	DefaultGasPriceDenom  = "uinit"
-	DefaultGasPriceAmount = "0.15"
-	InitiaTestnetChainId  = "initiation-2"
-	InitiaMainnetChainId  = "interwoven-1"
-
-	DefaultL1RelayerBalance = "1000000"
-	DefaultL2RelayerBalance = "1000000"
-	DefaultL1RelayerKeyName = "weave_l1_relayer"
-	DefaultL2RelayerKeyName = "weave_l2_relayer"
-)
